src/catalog/requirements.txt
flask
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-system-metrics
